+++ packages/aws-library/src/aws_library/s3/_constants.py
