+++ packages/aws-library/src/aws_library/s3/_error_handler.py
@@ -6,7 +6,7 @@ from typing import Any, Concatenate, ParamSpec, TypeVar
 
 from botocore import exceptions as botocore_exc
 
-from .errors import (
+from ._errors import (
     S3AccessError,
     S3BucketInvalidError,
     S3KeyNotFoundError,
