+++ packages/pytest-simcore/src/pytest_simcore/helpers/s3.py
@@ -1,11 +1,12 @@
+from collections.abc import Iterable
 from pathlib import Path
 from time import perf_counter
-from typing import Final, Iterable
+from typing import Final
 
 import aiofiles
 import orjson
 from aiohttp import ClientSession
-from aws_library.s3.models import MultiPartUploadLinks
+from aws_library.s3 import MultiPartUploadLinks
 from models_library.api_schemas_storage import ETag, FileUploadSchema, UploadedPart
 from pydantic import AnyUrl, ByteSize, parse_obj_as
 from servicelib.aiohttp import status
