+++ services/storage/src/simcore_service_storage/constants.py
@@ -1,6 +1,6 @@
 from typing import Final
 
-from aws_library.s3.constants import PRESIGNED_LINK_MAX_SIZE, S3_MAX_FILE_SIZE
+from aws_library.s3 import PRESIGNED_LINK_MAX_SIZE, S3_MAX_FILE_SIZE
 from models_library.api_schemas_storage import LinkType
 from pydantic import ByteSize
 from servicelib.aiohttp import application_keys
