+++ services/storage/src/simcore_service_storage/utils_handlers.py
@@ -3,7 +3,7 @@ import logging
 from aiohttp import web
 from aiohttp.typedefs import Handler
 from aiohttp.web_request import Request
-from aws_library.s3.errors import S3AccessError, S3KeyNotFoundError
+from aws_library.s3 import S3AccessError, S3KeyNotFoundError
 from pydantic import ValidationError
 from servicelib.aiohttp.aiopg_utils import DBAPIError
