+++ services/storage/tests/unit/test_handlers_simcore_s3.py
@@ -17,7 +17,7 @@ import sqlalchemy as sa
 from aiohttp import ClientResponseError
 from aiohttp.test_utils import TestClient
 from aiopg.sa.engine import Engine
-from aws_library.s3.client import SimcoreS3API
+from aws_library.s3 import SimcoreS3API
 from faker import Faker
 from models_library.api_schemas_storage import FileMetaDataGet, FoldersBody
 from models_library.basic_types import SHA256Str
