+++ services/storage/tests/unit/test_utils_handlers.py
@@ -6,7 +6,7 @@
 import pytest
 from aiohttp import web
 from aiohttp.typedefs import Handler
-from aws_library.s3.errors import S3KeyNotFoundError
+from aws_library.s3 import S3KeyNotFoundError
 from pydantic import BaseModel, ValidationError
 from pytest_mock import MockerFixture
 from servicelib.aiohttp.aiopg_utils import DBAPIError
